requirements.txt
Django
RPi.GPIO
requests
w1thermsensor
